Move student related views to the students app
