+++ home/templates/base.html
@@ -42,10 +42,10 @@
                         <a class="dropbtn">Teachers</a>
                         <div class="dropdown-content">
                             <a href="/teachers/">Teachers Home</a>
-                            <a href="/teachers/students/">View Students</a>
                             <a href="/commendations/award/">Award Commendation</a>
                             {% if request.user.teacher.is_management %}
-                                <a href="/teachers/badges/">Award spirit badges</a>
+                                <a href="/students/">View Students</a>
+                                <a href="/commendations/badges/">Award spirit badges</a>
                             {% endif %}
                         </div>
                     </div>
