+++ teachers/urls.py
@@ -4,7 +4,5 @@ from . import views
 
 urlpatterns = [
     path("", views.index, name="index"),
-    path("students/<int:ID>/", views.student, name="students"),
-    path("students/", views.students, name="student"),
     path("badges/", views.awardMilestones, name="badges"),
 ]
